requirements.txt
langchain
langchain_google_genai
gradio
huggingface_hub
python-dotenv
streamlit